Added current resume version
Currently considering reworking the Education field to draw more attention to the areas of study. Also considering removing the underlines of all the section titles, and adding another page for past work experience (despite not having any particularly relevant to a technical software position), and references. Formatting of the second page is up in the air.